+++ lib/queries/geographic_area/autocomplete.rb
@@ -27,6 +27,7 @@ module Queries
         queries = [
           ::GeographicArea.where(id: query_string).all,
           ::GeographicArea.where(name: query_string).all,
+          ::GeographicArea.joins(:alternate_values).where('alternate_values.value ILIKE ?', query_string).all,
           ::GeographicArea.joins(parent_child_join).where(Arel.sql(parent_child_where.to_sql)).limit(5).all,
           ::GeographicArea.where(Arel.sql(where_sql)).includes(:geographic_area_type, :geographic_items).order(Arel.sql('LENGTH(name)')).limit(dynamic_limit).all,
            autocomplete_exact_id,