+++ CHANGELOG.md
@@ -720,6 +720,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Recent predicate list
 - Recent languages list
 - People autocomplete
+- GeographicArea autocomplete (exact match on alternate value)
 - Remove search box in observation matrix hub [#3032]
 - Type material form allows multiple type species in comprehensive task. [#2584]
 - Updated Ruby gems.Yes